requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
PyMuPDF==1.23.8
python-docx==0.8.11
unstructured==0.11.2
ollama==0.1.7
chromadb==0.4.18
python-dotenv==1.0.0
loguru==0.7.2 